requirements.txt
pandas
python-Levenshtein
jellyfish